Move some code around, fix a couple of small things
